refactor: rename providers to contexts
Renamed `auth-provider` and `app-provider` to `auth-context` and `app-context` respectively for consistency with naming conventions. Updated all necessary imports and usage references for a seamless transition.
